@jbduncan 2024-03-13
@kortschak Here are my local benchmark results, running the benchmarks in [dade670](https://github.com/gonum/gonum/pull/1941/commits/dade6701d393f7b9b3f6c9267e851cd9f5f6ce32). (Previously I was using my Linux laptop but now I'm using my M1 MacBook.)

In summary, the difference between `master` and this PR looks to be insignificant.

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/cmplxs
                │ benchmarks_cmplxs_reverse_old.txt │ benchmarks_cmplxs_reverse_new.txt  │
                │              sec/op               │   sec/op     vs base               │
ReverseSmall-8                          5.322n ± 0%   5.321n ± 0%       ~ (p=0.576 n=10)
ReverseMedium-8                         397.6n ± 0%   397.9n ± 1%       ~ (p=0.696 n=10)
ReverseLarge-8                          44.26µ ± 1%   43.67µ ± 2%       ~ (p=0.063 n=10)
ReverseHuge-8                           5.335m ± 1%   5.337m ± 1%       ~ (p=0.971 n=10)
geomean                                 4.727µ        4.713µ       -0.30%
```