@jbduncan 2024-05-05
> Let's merge the last commit into the previous two (the deprecation into the commit that changes the function body and the whitespace change into the commit that adds the bechmarks).

Makes sense! Done. Ready for your next review.